Move 'move...' refactoring to RefactorProcessor
