+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/corrections/RefactorProcessor.java
@@ -18,8 +18,10 @@ import java.util.List;
 
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.jdt.core.dom.ASTNode;
+import org.eclipse.jdt.ls.core.internal.corrections.proposals.CUCorrectionProposal;
 import org.eclipse.jdt.ls.core.internal.corrections.proposals.ChangeCorrectionProposal;
 import org.eclipse.jdt.ls.core.internal.preferences.PreferenceManager;
+import org.eclipse.jdt.ls.core.internal.text.correction.RefactorProposalUtility;
 import org.eclipse.lsp4j.CodeActionParams;
 
 /**
@@ -40,6 +42,8 @@ public class RefactorProcessor {
 			// TODO (Yan): Move refactor proposals here.
 			InvertBooleanUtility.getInverseConditionProposals(params, context, coveringNode, proposals);
 			getInverseLocalVariableProposals(params, context, coveringNode, proposals);
+
+			getMoveRefactoringProposals(params, context, coveringNode, proposals);
 
 			return proposals;
 		}
@@ -64,6 +68,23 @@ public class RefactorProcessor {
 
 		proposals.add(proposal);
 		return true;
+	}
+
+	private boolean getMoveRefactoringProposals(CodeActionParams params, IInvocationContext context, ASTNode coveringNode, ArrayList<ChangeCorrectionProposal> resultingCollections) {
+		if (resultingCollections == null) {
+			return false;
+		}
+
+		if (this.preferenceManager.getClientPreferences().isMoveRefactoringSupported()) {
+			List<CUCorrectionProposal> newProposals = RefactorProposalUtility.getMoveRefactoringProposals(params, context);
+			if (newProposals != null && !newProposals.isEmpty()) {
+				resultingCollections.addAll(newProposals);
+				return true;
+			}
+		}
+
+		return false;
+
 	}
 
 }
